Refactor names, locations and remove unnecessary services
