+++ src/main/java/uk/gov/hmcts/cmc/claimstore/documents/output/PDF.java
@@ -1,8 +1,8 @@
-package uk.gov.hmcts.cmc.claimstore.documents;
+package uk.gov.hmcts.cmc.claimstore.documents.output;
 
 public class PDF {
     public static final String CONTENT_TYPE = "application/pdf";
-    public static final String EXTENSION = ".pdf";
+    private static final String EXTENSION = ".pdf";
 
     private final String filename;
     private final byte[] bytes;
